+++ docs/.vuepress/data/config.js
@@ -1,1 +1,1 @@
-export const publishDate = '2022-08-22';
+export const publishDate = '2022-09-22';